+++ tests/Query/QueryRendererHelper.test.ts
@@ -7,11 +7,15 @@ import { explainResults, getQueryForQueryRenderer } from '../../src/Query/QueryR
 import { GlobalFilter } from '../../src/Config/GlobalFilter';
 import { GlobalQuery } from '../../src/Config/GlobalQuery';
 import { TasksFile } from '../../src/Scripting/TasksFile';
+import { getTasksFileFromMockData } from '../TestingTools/MockDataHelpers';
+import query_file_defaults_short_mode from '../Obsidian/__test_data__/query_file_defaults_short_mode.json';
+import query_file_defaults_all_options_true from '../Obsidian/__test_data__/query_file_defaults_all_options_true.json';
+import query_file_defaults_ignore_global_query from '../Obsidian/__test_data__/query_file_defaults_ignore_global_query.json';
 
 window.moment = moment;
 
 describe('explain', () => {
-    it('should explain a task', () => {
+    it('should explain a search', () => {
         const source = '';
         const query = new Query(source);
         expect(explainResults(query.source, new GlobalFilter(), new GlobalQuery())).toMatchInlineSnapshot(`
@@ -26,7 +30,7 @@ describe('explain', () => {
         `);
     });
 
-    it('should explain a task with global filter active', () => {
+    it('should explain a search with global filter active', () => {
         const globalFilter = new GlobalFilter();
         globalFilter.set('#task');
 
@@ -46,7 +50,7 @@ describe('explain', () => {
         `);
     });
 
-    it('should explain a task with global query active', () => {
+    it('should explain a search with global query active', () => {
         const globalQuery = new GlobalQuery('description includes hello');
 
         const source = '';
@@ -71,7 +75,41 @@ describe('explain', () => {
         `);
     });
 
-    it('should explain a task with global query and global filter active', () => {
+    it('should explain a search with global query and query file defaults active', () => {
+        const globalQuery = new GlobalQuery('description includes hello');
+
+        const source = '';
+        const query = new Query(source);
+        const tasksFile = getTasksFileFromMockData(query_file_defaults_all_options_true);
+        expect(explainResults(query.source, new GlobalFilter(), globalQuery, tasksFile)).toMatchInlineSnapshot(`
+            "Explanation of the global query:
+
+              description includes hello
+
+              No grouping instructions supplied.
+
+              No sorting instructions supplied.
+
+            Explanation of the query file defaults (from properties/frontmatter in the query's file):
+
+              not done
+
+              No grouping instructions supplied.
+
+              No sorting instructions supplied.
+
+            Explanation of this Tasks code block query:
+
+              No filters supplied. All tasks will match the query.
+
+              No grouping instructions supplied.
+
+              No sorting instructions supplied.
+            "
+        `);
+    });
+
+    it('should explain a search with global query and global filter active', () => {
         const globalQuery = new GlobalQuery('description includes hello');
         const globalFilter = new GlobalFilter();
         globalFilter.set('#task');
@@ -100,7 +138,7 @@ describe('explain', () => {
         `);
     });
 
-    it('should explain a task with global query set but ignored without the global query', () => {
+    it('should explain a search with global query set but ignored without the global query', () => {
         const globalQuery = new GlobalQuery('description includes hello');
 
         const source = 'ignore global query';
@@ -109,6 +147,33 @@ describe('explain', () => {
             "Explanation of this Tasks code block query:
 
               No filters supplied. All tasks will match the query.
+
+              No grouping instructions supplied.
+
+              No sorting instructions supplied.
+            "
+        `);
+    });
+
+    it('should explain a search with global query set but ignored via tasks_query_extra_instructions saying to ignore it', () => {
+        const globalQuery = new GlobalQuery('description includes hello');
+
+        const source = 'description includes I came from the code block';
+        const queryFile = getTasksFileFromMockData(query_file_defaults_ignore_global_query);
+        expect(queryFile.property('tasks_query_extra_instructions')).toContain('ignore global query');
+        const query = new Query(source, queryFile);
+        expect(explainResults(query.source, new GlobalFilter(), globalQuery, queryFile)).toMatchInlineSnapshot(`
+            "Explanation of the query file defaults (from properties/frontmatter in the query's file):
+
+              description includes I came from the tasks_query_extra_instructions property
+
+              No grouping instructions supplied.
+
+              No sorting instructions supplied.
+
+            Explanation of this Tasks code block query:
+
+              description includes I came from the code block
 
               No grouping instructions supplied.
 
@@ -153,6 +218,21 @@ describe('query used for QueryRenderer', () => {
 
         // Assert
         expect(query.source).toEqual('description includes from_block_query\nignore global query');
+        expect(query.tasksFile).toBe(tasksFile);
+    });
+
+    it('should add QueryFileDefaults', () => {
+        // Arrange
+        const globalQuery = new GlobalQuery('path includes from_global_query');
+        const tasksFile = getTasksFileFromMockData(query_file_defaults_short_mode);
+
+        // Act
+        const query = getQueryForQueryRenderer('description includes from_block_query', globalQuery, tasksFile);
+
+        // Assert
+        expect(query.source).toEqual(`path includes from_global_query
+short mode
+description includes from_block_query`);
         expect(query.tasksFile).toBe(tasksFile);
     });
 });