+++ docs/What is New/Changelog.md
@@ -14,8 +14,10 @@ _In recent [Tasks releases](https://github.com/obsidian-tasks-group/obsidian-tas
 
 - X.Y.Z:
   - Add `query.file.hasProperty()` and `query.file.property()` in custom filters
-  - Add `{{query.file.hasProperty()}}` and `{{query.file.property()}}` in placeholders - see [[Obsidian Properties#Using Query Properties in Placeholders|Using Query Properties in Placeholders]].
+  - Add `{{query.file.hasProperty()}}` and `{{query.file.property()}}` in placeholders - see [[Obsidian Properties#Using Query Properties in Searches|Using Query Properties in Placeholders]].
   - Placeholders can now call functions and contain expressions.
+  - New [[Query File Defaults]] facility allows easy customising of all the
+    Tasks searches in a file, by editing properties
   - Add Chinese translation of [[Settings]], [[Editing a Status]] and [[Check your Statuses]]
 - 7.14.0:
   - Add [[Editing Dates#Date-picker on task dates|date picker]] to Reading mode and Tasks query search results.